(pub) [refactoring] Home repo structure
