+++ src/views/Home/atoms/CardRank.vue
